+++ src/cjs.js
@@ -1,1 +1,3 @@
 module.exports = require('./index').default;
+module.exports.storageSingleton = require('./storage-singleton').default;
+module.exports.assetsManifestPlugin = require('./assets-manifest-plugin').default;